backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
asgi-lifespan==2.1.0
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
beanie==1.20.0
boto3==1.28.1
botocore==1.31.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
coverage==7.2.7
cryptography==41.0.1
dnspython==2.3.0
ecdsa==0.18.0
exceptiongroup==1.1.2
fastapi==0.100.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
jmespath==1.0.1
lazy-model==0.0.5
mongomock==4.1.2
mongomock-motor==0.0.20
motor==3.2.0
multidict==6.0.4
openai==0.27.9
outcome==1.2.0
packaging==23.1
passlib==1.7.4
pluggy==1.2.0
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.11
pydantic_core==2.1.2
PyJWT==2.8.0
pymongo==4.4.0
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0
pytest-mock==3.11.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.31.0
rsa==4.9
s3transfer==0.6.1
sentinels==1.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.27.0
stripe==6.2.0
toml==0.10.2
tomli==2.0.1
tqdm==4.65.0
trio==0.22.1
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
